Remove chpl populator package as it is no longer necessary and put everything back in main file, fix product querier structure that endpoint list product file is unmarshalled into
